+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 spring.servlet.multipart.max-file-size=100MB
 spring.servlet.multipart.max-request-size=100MB
 audio.video.path=thirdParty/ffmpeg/ffmpeg.exe
-image.path=thirdParty/imagemagic/image.exe
+image.path=/imthirdPartyagemagic/image.exe
 metadata.path=thirdParty/exiftool/exiftool.exe
-ocr.path=thirdParty/ocr/ocr.exe
+ocr.path=thirdParty/ocr/tesseract.exe
 convertedFiles.path=storage/convertedFiles/
 inputFiles.path=storage/inputFiles/
